pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "snapshotter"
version = "1.0.0"
description = "Unified backup orchestration service for multi-target infrastructure"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Lesaloon", email = "contact@example.com"}
]
keywords = [
    "backup",
    "orchestration",
    "postgresql",
    "prometheus",
    "snapshots"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Archiving :: Backup",
]

dependencies = [
    "PyYAML>=6.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.10",
]

[project.scripts]
snapshotter = "snapshotter.__main__:main"

[tool.setuptools]
packages = ["snapshotter", "snapshotter.backups", "snapshotter.notifiers", "snapshotter.utils"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "W"]
exclude = [".git", "__pycache__", "venv", ".venv"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=snapshotter --cov-report=term-missing --cov-report=html"
